FlightService/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django~=4.1.1
django-choices==1.7.2
djangorestframework~=3.13.1
idna==3.4
psycopg2-binary==2.9.4
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse~=0.4.2
typing_extensions==4.4.0
urllib3==1.26.12

pip~=22.3
keyring~=18.0.1
wheel~=0.34.2
psycopg2~=2.8.4
colorama~=0.4.3
PyYAML~=5.3.1
Jinja2~=3.0.1
numpy~=1.23.3
cryptography~=2.8
chardet~=3.0.4
simplejson~=3.16.0
setuptools~=45.2.0
jsonschema~=3.2.0
circuitbreaker~=1.4.0
kafka~=1.3.5
future~=0.18.2
amqp~=5.1.1
vine~=5.0.0
kombu~=5.2.4
billiard~=3.6.4.0
redis~=4.4.0
celery~=5.2.7
click~=8.1.3
greenlet~=1.1.3
pytest~=7.2.0
gevent~=21.12.0
wcwidth~=0.2.5